Sub-conjunto CASEN 2017

Datos de demostracion limitados a las comunas de la Region de Los Rios con representatividad comunal (Valdivia, Los Rios, Paillaco, Panguipulli, La Union y Rio Bueno).

casen_2017_los_rios

Format

Un data frame con 8055 observaciones en las siguientes 7 variables.

expc

factor de expansion comunal

varstrat

estratos de varianza

varunit

conglomerados de varianza

ytotcorh

ingreso total del hogar corregido

sexo

sexo

comuna

comuna

pobreza

situacion de pobreza por ingresos

Source

Elaboracion propia con base en datos de Ministerio de Desarrollo Social (MDS).

Examples

casen_2017_los_rios
#> # A tibble: 8,055 × 7
#>     expc varstrat     varunit ytotcorh       sexo           comuna       pobreza
#>    <dbl>    <dbl>       <dbl>    <dbl>  <dbl+lbl>        <dbl+lbl>     <dbl+lbl>
#>  1    53   141283 14128301184   959765 2 [Mujer]  14101 [Valdivia] 3 [No pobres]
#>  2    53   141283 14128301184   959765 2 [Mujer]  14101 [Valdivia] 3 [No pobres]
#>  3    53   141283 14128301184   959765 2 [Mujer]  14101 [Valdivia] 3 [No pobres]
#>  4   105   141283 14128301184   283273 2 [Mujer]  14101 [Valdivia] 2 [Pobres no…
#>  5   105   141283 14128301184   283273 1 [Hombre] 14101 [Valdivia] 2 [Pobres no…
#>  6   105   141283 14128301184   283273 2 [Mujer]  14101 [Valdivia] 2 [Pobres no…
#>  7   105   141283 14128301184   353083 2 [Mujer]  14101 [Valdivia] 3 [No pobres]
#>  8   105   141283 14128301184   260000 2 [Mujer]  14101 [Valdivia] 2 [Pobres no…
#>  9   105   141283 14128301184   260000 2 [Mujer]  14101 [Valdivia] 2 [Pobres no…
#> 10   105   141283 14128301184   260000 1 [Hombre] 14101 [Valdivia] 2 [Pobres no…
#> # … with 8,045 more rows