Lee los archivos descargados de la encuesta CASEN

Lee los archivos descargados de la encuesta CASEN

leer_casen(archivo)

Arguments

archivo
cadena de texto con la ruta de la encuesta en formato zip/rar o sav/dta

Value

la encuesta CASEN en formato tibble

Examples

# datos de ejemplo en el paquete
leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))

#> Intentando leer el archivo /tmp/Rtmp0WoxBB/temp_libpath431f4b0542e1/casen/extdata/casen_2017_los_rios.zip...

#> Los metadatos de /tmp/Rtmp0WoxBB/temp_libpath431f4b0542e1/casen/extdata/casen_2017_los_rios.zip indican que el formato comprimido es adecuado, intentando leer...

#> Resizing buffer to 32768.
#> Resizing buffer to 65536.
#> Resizing buffer to 131072.
#> Resizing buffer to 262144.

#> /tmp/Rtmp0WoxBB/temp_libpath431f4b0542e1/casen/extdata/casen_2017_los_rios.zip se pudo leer como tibble :-)

#> # A tibble: 10,172 x 8
#>     expr  expc varstrat   varunit ytotcorh      sexo        comuna       pobreza
#>    <dbl> <dbl>    <dbl>     <dbl>    <dbl> <dbl+lbl>     <dbl+lbl>     <dbl+lbl>
#>  1    50    53   141283   1.41e10   959765 2 [Mujer] 14101 [Valdi… 3 [No pobres]
#>  2    50    53   141283   1.41e10   959765 2 [Mujer] 14101 [Valdi… 3 [No pobres]
#>  3    50    53   141283   1.41e10   959765 2 [Mujer] 14101 [Valdi… 3 [No pobres]
#>  4   100   105   141283   1.41e10   283273 2 [Mujer] 14101 [Valdi… 2 [Pobres no…
#>  5   100   105   141283   1.41e10   283273 1 [Hombr… 14101 [Valdi… 2 [Pobres no…
#>  6   100   105   141283   1.41e10   283273 2 [Mujer] 14101 [Valdi… 2 [Pobres no…
#>  7   100   105   141283   1.41e10   353083 2 [Mujer] 14101 [Valdi… 3 [No pobres]
#>  8   100   105   141283   1.41e10   260000 2 [Mujer] 14101 [Valdi… 2 [Pobres no…
#>  9   100   105   141283   1.41e10   260000 2 [Mujer] 14101 [Valdi… 2 [Pobres no…
#> 10   100   105   141283   1.41e10   260000 1 [Hombr… 14101 [Valdi… 2 [Pobres no…
#> # … with 10,162 more rows

# descargando los datos
# descargar_casen(2017, "data-raw")
# leer_casen("data-raw/2017_spss.rar")