Media agrupada usando diseño complejo

Usa los factores de expansion, conglomerados y estratos para calcular correctamente las medias comunales o regionales.

media_agrupada(disenio)

Arguments

disenio
la salida de `configuracion_disenio()` que provee ademas los grupos y las variables en forma de lista

Value

Una tabla con las medias agrupadas y su intervalo de confianza.

Examples

cd <- configuracion_disenio(casen_2017_los_rios, "ytotcorh", c("comuna", "sexo"), "expc")
media_agrupada(cd)

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> `mutate_if()` ignored the following grouping variables:
#> Column `comuna`

#> # A tibble: 12 x 8
#> # Groups:   comuna_etiqueta [6]
#>    comuna_etiqueta sexo_etiqueta comuna_codigo sexo_codigo      comuna
#>          <dbl+lbl> <chr>             <dbl+lbl> <chr>         <dbl+lbl>
#>  1 14101 [Valdivi… Mujer         14101 [Valdi… 2           14101 [Val…
#>  2 14101 [Valdivi… Hombre        14101 [Valdi… 1           14101 [Val…
#>  3 14104 [Los Lag… Hombre        14104 [Los L… 1           14104 [Los…
#>  4 14104 [Los Lag… Mujer         14104 [Los L… 2           14104 [Los…
#>  5 14107 [Paillac… Hombre        14107 [Paill… 1           14107 [Pai…
#>  6 14107 [Paillac… Mujer         14107 [Paill… 2           14107 [Pai…
#>  7 14108 [Panguip… Mujer         14108 [Pangu… 2           14108 [Pan…
#>  8 14108 [Panguip… Hombre        14108 [Pangu… 1           14108 [Pan…
#>  9 14201 [La Unió… Hombre        14201 [La Un… 1           14201 [La …
#> 10 14201 [La Unió… Mujer         14201 [La Un… 2           14201 [La …
#> 11 14204 [Río Bue… Hombre        14204 [Río B… 1           14204 [Río…
#> 12 14204 [Río Bue… Mujer         14204 [Río B… 2           14204 [Río…
#> # … with 3 more variables: media_ytotcorh <dbl>, media_ytotcorh_inf <dbl>,
#> #   media_ytotcorh_sup <dbl>