Modelo lineal generalizado con disenio complejo

Modelo lineal generalizado con disenio complejo

modelo_lineal_generalizado(datos, variable_dependiente = "ytotcorh",
  variables_independientes = "region", peso = "expr",
  conglomerado = "varunit", estrato = "varstrat")

Arguments

datos
un data.frame o tibble con la encuesta CASEN (o un subconjunto acotado a una region, etc)

variable_dependiente
una columna de tipo numerico, por ejemplo ytotcorh que es la opcion por defecto

variables_independientes
una o mas columnas de tipo numerico o factor, por ejemplo region que es la opcion por defecto

peso
una columna de tipo numerico, por defecto es expr que corresponde al factor de expansion regional de acuerdo al manual CASEN 2017

conglomerado
una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017

estrato
una columna de tipo numerico, por defecto es varunit de acuerdo al manual CASEN 2017

Value

un tibble con los resultados del modelo lineal generalizado (funcion glm + disenio complejo)

Examples

r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))

#> Intentando leer el archivo /tmp/RtmpV5bXwB/temp_libpath67fa463e0a6/casen/extdata/casen_2017_los_rios.zip...

#> Los metadatos de /tmp/RtmpV5bXwB/temp_libpath67fa463e0a6/casen/extdata/casen_2017_los_rios.zip indican que el formato comprimido es adecuado, intentando leer...

#> Resizing buffer to 32768.
#> Resizing buffer to 65536.
#> Resizing buffer to 131072.
#> Resizing buffer to 262144.

#> /tmp/RtmpV5bXwB/temp_libpath67fa463e0a6/casen/extdata/casen_2017_los_rios.zip se pudo leer como tibble :-)

modelo_lineal_generalizado(r14, "ytotcorh", c("comuna", "sexo"), "expc")

#> # A tibble: 13 x 7
#>    term              estimate std_error statistic  p_value conf_low conf_high
#>    <chr>                <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#>  1 (Intercept)       1331959.   128517.     10.4  1.53e-14 1075438.  1588481.
#>  2 comunaCorral      -642591.   137144.     -4.69 1.88e- 5 -916333.  -368850.
#>  3 comunaLanco       -621449.   132259.     -4.70 1.79e- 5 -885439.  -357460.
#>  4 comunaLos Lagos   -579666.   125585.     -4.62 2.39e- 5 -830335.  -328998.
#>  5 comunaMáfil       -465695.   134021.     -3.47 1.00e- 3 -733203.  -198188.
#>  6 comunaMariquina   -532944.   163291.     -3.26 1.89e- 3 -858874.  -207013.
#>  7 comunaPaillaco    -605114.   135751.     -4.46 4.12e- 5 -876074.  -334154.
#>  8 comunaPanguipulli -645983.   125267.     -5.16 3.53e- 6 -896017.  -395949.
#>  9 comunaLa Unión    -525886.   128681.     -4.09 1.43e- 4 -782734.  -269038.
#> 10 comunaFutrono     -532511.   136537.     -3.90 2.64e- 4 -805040.  -259982.
#> 11 comunaLago Ranco  -584849.   130485.     -4.48 3.79e- 5 -845299.  -324399.
#> 12 comunaRío Bueno   -560813.   130805.     -4.29 7.35e- 5 -821901.  -299726.
#> 13 sexoMujer          -55931.    20067.     -2.79 7.28e- 3  -95984.   -15878.