Percentiles agrupados usando diseño complejo

Usa los factores de expansion, conglomerados y estratos para calcular correctamente los percentiles comunales o regionales.

percentiles_agrupados(disenio, percentiles = 0.7)

Arguments

disenio

la salida de `configuracion_disenio()` que provee ademas los grupos y las variables en forma de lista

percentiles

percentiles a calcular, si no se especifica calcula el percentil 70

Value

Una tabla con los percentiles y su error estandar.

Examples

cd <- configuracion_disenio(casen_2017_los_rios, "ytotcorh", c("comuna", "sexo"), "expc")
percentiles_agrupados(cd, 0.7)
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> `mutate_if()` ignored the following grouping variables:
#> • Column `comuna`
#> # A tibble: 12 × 8
#> # Groups:   comuna_etiqueta [6]
#>    percentil comuna_etiqueta sexo_etiqueta comuna_codigo sexo_codigo      comuna
#>        <dbl>       <dbl+lbl> <chr>             <dbl+lbl> <chr>         <dbl+lbl>
#>  1       0.7 14101 [Valdivi… Mujer         14101 [Valdi… 2           14101 [Val…
#>  2       0.7 14101 [Valdivi… Hombre        14101 [Valdi… 1           14101 [Val…
#>  3       0.7 14104 [Los Lag… Hombre        14104 [Los L… 1           14104 [Los…
#>  4       0.7 14104 [Los Lag… Mujer         14104 [Los L… 2           14104 [Los…
#>  5       0.7 14107 [Paillac… Hombre        14107 [Paill… 1           14107 [Pai…
#>  6       0.7 14107 [Paillac… Mujer         14107 [Paill… 2           14107 [Pai…
#>  7       0.7 14108 [Panguip… Mujer         14108 [Pangu… 2           14108 [Pan…
#>  8       0.7 14108 [Panguip… Hombre        14108 [Pangu… 1           14108 [Pan…
#>  9       0.7 14201 [La Unió… Hombre        14201 [La Un… 1           14201 [La …
#> 10       0.7 14201 [La Unió… Mujer         14201 [La Un… 2           14201 [La …
#> 11       0.7 14204 [Río Bue… Hombre        14204 [Río B… 1           14204 [Río…
#> 12       0.7 14204 [Río Bue… Mujer         14204 [Río B… 2           14204 [Río…
#> # … with 2 more variables: mediana_ytotcorh <dbl>,
#> #   mediana_ytotcorh_err_est <dbl>