Modelo lineal generalizado con disenio complejo

modelo_lineal_generalizado define el disenio a partir de los conglomerados, con base en esto obtiene ajusta un modelo lineal usando minimos cuadrados generalizados.

modelo_lineal_generalizado(disenio, modelo = "ytotcorh ~ sexo")

Arguments

disenio
la salida de `configuracion_disenio()` que provee ademas los grupos y las variables en forma de lista

modelo
una expresion sinmbolica escrita como cadena de texto, consulta [stats::formula()] para los detalles de modelos con componentes logaritmicas, cuadraticas, etc.

Value

una lista cuyas clases son svyglm, glm y lm

Examples

r14 <- leer_casen(system.file(package = "casen", "extdata", "casen_2017_los_rios.zip"))

#> Intentando leer el archivo /tmp/RtmpMqJAvD/temp_libpath5ca48a3ad38/casen/extdata/casen_2017_los_rios.zip...

#> Los metadatos de /tmp/RtmpMqJAvD/temp_libpath5ca48a3ad38/casen/extdata/casen_2017_los_rios.zip indican que el formato comprimido es adecuado, intentando leer...

#> Resizing buffer to 32768.
#> Resizing buffer to 65536.
#> Resizing buffer to 131072.
#> Resizing buffer to 262144.

#> /tmp/RtmpMqJAvD/temp_libpath5ca48a3ad38/casen/extdata/casen_2017_los_rios.zip se pudo leer como tibble :-)

cd <- configuracion_disenio(r14, "ytotcorh", c("comuna", "sexo"), "expc")
modelo_lineal_generalizado(cd, "ytotcorh ~ comuna + sexo")

#> Stratified 1 - level Cluster Sampling design (with replacement)
#> With (83) clusters.
#> Called via srvyr
#> Sampling variables:
#>  - ids: varunit
#>  - strata: varstrat
#>  - weights: expc
#> 
#> Call:  svyglm(formula = stats::as.formula(modelo), design = disenio$disenio)
#> 
#> Coefficients:
#> (Intercept)       comuna         sexo  
#>    47040565        -3255       -49948  
#> 
#> Degrees of Freedom: 10171 Total (i.e. Null);  65 Residual
#> Null Deviance:	    1.319e+16 
#> Residual Deviance: 1.299e+16 	AIC: 313600